refactoring for server.js
