+++ gulp/build.js
@@ -5,34 +5,16 @@ const tsify = require('tsify');
 const uglify = require('gulp-uglify');
 const sourcemaps = require('gulp-sourcemaps');
 const buffer = require('vinyl-buffer');
-const tslint = require('gulp-tslint');
-const eslint = require('gulp-eslint');
-const connect = require('gulp-connect');
 const gutil = require('gulp-util');
 const runSequence = require('run-sequence');
 const staticFiles = require('./staticFiles');
 const tests = require('./tests');
 const clean = require('./clean');
-
-const paths = {
-  pages: ['../src/*.html'],
-  jsSrcs: ['./**/*.js'],
-  tsSrcs: ['../src/**/*.ts'],
-};
+require('./lint');
 
 gulp.task('clean-build', (done) => {
   clean.run(done, './build');
 });
-
-gulp.task('eslint', () => gulp.src(paths.jsSrcs)
-    .pipe(eslint())
-    .pipe(eslint.format()));
-
-gulp.task('tslint', () => gulp.src(paths.tsSrcs)
-    .pipe(tslint({
-      formatter: 'verbose',
-    }))
-    .pipe(tslint.report()));
 
 const buildBrowserify = browserify({
   basedir: '.',
@@ -62,5 +44,5 @@ function build() {
 // todo: add delete path
 gulp.task('build', () => {
   runSequence(['eslint', 'tslint']);
-  build();
+  return build();
 });
