--- index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="UTF-8" />
-        <title>Boilerplate!</title>
-    </head>
-    <body>
-        <p id="greeting">Loading ...</p>
-        <script src="index.js"></script>
-    </body>
-</html>
